make alliance page
