+++ client/models/FIRST/team.ts
@@ -266,16 +266,15 @@ export class FIRSTTeam extends Cache<FIRSTTeamEventData> {
 
             if (res.isErr()) throw res.error;
 
-            return res.value
-                .filter((p, i, a) => {
-                    return (
-                        a.findIndex(
-                            p2 =>
-                                p2.matchNumber === p.matchNumber &&
-                                p2.compLevel === p.compLevel
-                        ) === i
-                    );
-                });
+            return res.value.filter((p, i, a) => {
+                return (
+                    a.findIndex(
+                        p2 =>
+                            p2.matchNumber === p.matchNumber &&
+                            p2.compLevel === p.compLevel
+                    ) === i
+                );
+            });
         });
     }
 
@@ -387,12 +386,13 @@ export class FIRSTTeam extends Cache<FIRSTTeamEventData> {
 
             if (res.isErr()) throw res.error;
             return res.value.filter(
-                (s, i, a) => a.findIndex(_s => {
-                    return (
-                        _s.matchNumber === s.matchNumber &&
-                        _s.compLevel === s.compLevel
-                    );
-                }) === i
+                (s, i, a) =>
+                    a.findIndex(_s => {
+                        return (
+                            _s.matchNumber === s.matchNumber &&
+                            _s.compLevel === s.compLevel
+                        );
+                    }) === i
             );
         });
     }
