+++ client/views/components/robot-display/MatchesSummaryChart.svelte
@@ -29,7 +29,7 @@ const fns = {
         data = {
             labels: [],
             datasets: []
-        }
+        };
 
         const matches = await team.getMatchScouting();
