+++ public/templates/index.html
@@ -42,7 +42,10 @@
             referrerpolicy="no-referrer"
         />
 
-        <link href="https://fonts.googleapis.com/css2?family=Material+Symbols+Outlined" rel="stylesheet" />
+        <link
+            href="https://fonts.googleapis.com/css2?family=Material+Symbols+Outlined"
+            rel="stylesheet"
+        />
 
         <link rel="stylesheet" href="{{ style }}" />
     </head>
